# Add a fast path for adding new node in node_authorizer