**What type of PR is this?**
/kind feature

**What this PR does / why we need it**:
This is a performance improvement for scheduling new pod with secrets, configmaps and other resources.
Without this patch, each pod using secret X will trigger index recomputation of size O(pods using that secret), so in clusters with highly replicated secrets this can be a huge cost.
With this patch, each such pod will be handled in O(1) time.

We saw that this was impacting scalability tests when we created pods with secrets with high rate (100 pods/s), creating a O(minute) delay between the time when pod was scheduled on some node and time when node was able to fetch associated secret.

Similar improvement can be implemented for removing secrets, but this will require introducing reference counting, so maybe in future.


This seems to improve WriteIndexMaintenance benchmark:
```
Before:
BenchmarkWriteIndexMaintenance-12    	    1034	   1157922 ns/op	    1906 B/op	      41 allocs/op
After:
BenchmarkWriteIndexMaintenance-12    	    4891	    239821 ns/op	    1572 B/op	      37 allocs/op
```

This also changes the way the index is used in node_authorizer: before this patch, even if destinationEdgeIndex is present for given node, but there is no path between two nodes, DFS will be triggered. This can be simplified, as if destinationEdgeIndex is present, it must be complete so we can easily fail if destinationEdgeIndex is present, but destination node is not there.

**Which issue(s) this PR fixes**:
<!--
*Automatically closes linked issue when PR is merged.
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
_If PR is about `failing-tests or flakes`, please post the related issues/tests in a comment and do not use `Fixes`_*
-->
Fixes #

**Special notes for your reviewer**:

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
If yes, a release note is required:
Enter your extended release note in the block below. If the PR requires additional action from users switching to the new release, include the string "action required".

For more information on release notes see: https://git.k8s.io/community/contributors/guide/release-notes.md
-->
```release-note
NONE
```

**Additional documentation e.g., KEPs (Kubernetes Enhancement Proposals), usage docs, etc.**:

<!--
This section can be blank if this pull request does not require a release note.

When adding links which point to resources within git repositories, like
KEPs or supporting documentation, please reference a specific commit and avoid
linking directly to the master branch. This ensures that links reference a
specific point in time, rather than a document that may change over time.

See here for guidance on getting permanent links to files: https://help.github.com/en/articles/getting-permanent-links-to-files

Please use the following format for linking documentation:
- [KEP]: <link>
- [Usage]: <link>
- [Other doc]: <link>
-->
```docs

```

/assign @wojtek-t 
/assign @liggitt