@liggitt 2020-01-30
While reviewing this, I found an existing bug in the index maintenance code. Opened https://github.com/kubernetes/kubernetes/pull/87693 to resolve that first.