@liggitt 2020-01-30
opened WIP https://github.com/kubernetes/kubernetes/pull/87696 that builds on this and switches to refcount and adds a fastpath for edge removal.
```
benchmark                             old ns/op     new ns/op     delta
BenchmarkWriteIndexMaintenance-12     1935450       8629          -99.55%

benchmark                             old allocs     new allocs     delta
BenchmarkWriteIndexMaintenance-12     44             37             -15.91%

benchmark                             old bytes     new bytes     delta
BenchmarkWriteIndexMaintenance-12     2141          1521          -28.96%
```