@liggitt 2020-01-31
/retest
/test pull-kubernetes-bazel-test