@liggitt 2020-01-31
/hold

the destination index is not authoritative, since not all edges know their destination node. see https://github.com/kubernetes/kubernetes/pull/87688#discussion_r373337027