@liggitt 2020-01-31
the fastpath index add is useful and should be kept (and the corresponding refcount and fastpath remove in https://github.com/kubernetes/kubernetes/pull/87696)