@mborsz 2020-01-31
Thanks for finding the issue with destination index. I reverted this part.